Move Post, PostList to separate files
