+++ src/js/components/Home.vue
@@ -5,7 +5,6 @@
 <script>
 import PostList from './PostList.vue';
 export default {
-    name: 'home',
     components: {
         PostList
     }
