+++ src/js/router/index.js
@@ -1,17 +1,18 @@
 import Vue from 'vue';
 import Router from 'vue-router';
 import Home from '../components/Home.vue'
+import Post from '../components/Post.vue'
 
 Vue.use(Router);
 
 export default new Router({
     routes: [
-        { path: '/', component: Home, name: 'Home' }
-        // {
-        //     path: '/post',
-        //     name: 'post',
-        //     component: Post
-        // },
+        { path: '/', component: Home, name: 'Home' },
+        {
+            path: '/post',
+            name: 'post',
+            component: Post
+        }
         // {
         //     path: '/comments',
         //     name: 'comments',
